DPR-531 move field filters into separate package
